requirements.txt
streamlit
Pillow
pymupdf
google-generativeai
google-cloud-texttospeech